requirements.txt
pandas>=2.0.0
openpyxl>=3.0.0
twilio>=8.0.0
APScheduler>=3.10.0
python-dotenv>=1.0.0
phonenumbers>=8.13.0
pyyaml>=6.0

